refactor postgres query

+++ app/lib/actions.ts
@@ -140,9 +140,9 @@ export async function addToFavourites(user_id: string, artist_id: string, artist
   }
 }
 
-export async function getFavourite(artistId: string) {
+export async function getFavourite(artistId: string, userId: string) {
   const favourite =
-    await sql`SELECT * FROM favourites WHERE artist_id=${artistId}`;
+    await sql`SELECT * FROM favourites WHERE artist_id = ${artistId} AND user_id = ${userId}`;
   
   if (favourite) {
     return true;
@@ -151,9 +151,9 @@ export async function getFavourite(artistId: string) {
   }
 }
 
-export async function deleteFavourite(artistId: string) {
+export async function deleteFavourite(artistId: string, userId: string) {
   const favourite =
-    await sql`DELETE FROM favourites WHERE artist_id=${artistId}`;
+    await sql`DELETE FROM favourites WHERE artist_id=${artistId} AND user_id = ${userId}`;
   
   return true;
 }
